+++ crates/module-system/sov-test-utils/src/interface/roles/mod.rs
@@ -20,7 +20,7 @@ pub struct UserTokenInfo {
     /// The associated token name
     pub token_name: TestTokenName,
     /// The user balance
-    pub balance: u64,
+    pub balance: u128,
     /// If the user can mint the token
     pub is_minter: bool,
 }
@@ -31,7 +31,7 @@ pub struct TestUser<S: Spec> {
     /// The private key of the user.
     pub private_key: <S::CryptoSpec as CryptoSpec>::PrivateKey,
     /// The bank balance of the user for the default gas token.
-    pub available_gas_balance: u64,
+    pub available_gas_balance: u128,
     /// The balances of the user for each non-gas token.
     pub token_balances: Vec<UserTokenInfo>,
     /// A custom credential id that can be defined at genesis.
@@ -40,7 +40,7 @@ pub struct TestUser<S: Spec> {
 
 impl<S: Spec> TestUser<S> {
     /// Creates a new user with the given private key and balance.
-    pub fn new(private_key: <S::CryptoSpec as CryptoSpec>::PrivateKey, balance: u64) -> Self {
+    pub fn new(private_key: <S::CryptoSpec as CryptoSpec>::PrivateKey, balance: u128) -> Self {
         Self {
             private_key,
             available_gas_balance: balance,
@@ -55,7 +55,7 @@ impl<S: Spec> TestUser<S> {
     }
 
     /// Generates a new user with the given balance.
-    pub fn generate(balance: u64) -> Self {
+    pub fn generate(balance: u128) -> Self {
         Self {
             private_key: <<S as Spec>::CryptoSpec as CryptoSpec>::PrivateKey::generate(),
             available_gas_balance: balance,
@@ -100,12 +100,12 @@ impl<S: Spec> TestUser<S> {
     }
 
     /// Returns the balance of the user.
-    pub fn balance(&self) -> u64 {
+    pub fn balance(&self) -> u128 {
         self.available_gas_balance
     }
 
     /// Returns the balance of the user for the given token.
-    pub fn token_balance(&self, token_name: &TestTokenName) -> Option<u64> {
+    pub fn token_balance(&self, token_name: &TestTokenName) -> Option<u128> {
         self.token_balances
             .iter()
             .find(|info| info.token_name == *token_name)
